+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## Unreleased
+
+#### 🚀 Updates
+
+- Updated remote caching to error if an output file is a symlink to a file outside of the workspace.
+
+#### 🐞 Fixes
+
+- Fixed an issue with priority tasks where dependencies of the task may sometimes not run in the
+  correct order in the pipeline.
+- Fixed an issue where task `affectedFiles` and `runFromWorkspaceRoot` would pass invalid relative
+  file paths.
+- Fixed an issue where running tasks for custom plugins would trigger a panic.
+
 ## 1.35.3
 
 #### 🚀 Updates